Separate out EnkfObs from EnKFMain
