+++ tests/unit_tests/c_wrappers/res/enkf/test_enkf_obs.py
@@ -5,7 +5,13 @@ from textwrap import dedent
 import pytest
 from ecl.summary import EclSum
 
-from ert._c_wrappers.enkf import EnkfObs, ErtConfig
+from ert._c_wrappers.enkf import EnkfObs, ErtConfig, ObsVector
+from ert._c_wrappers.enkf.config import EnkfConfigNode
+from ert._c_wrappers.enkf.enums import (
+    EnkfObservationImplementationType,
+    LoadFailTypeEnum,
+)
+from ert._c_wrappers.enkf.observations.summary_observation import SummaryObservation
 
 
 def run_simulator():
@@ -117,3 +123,52 @@ def test_date_parsing_in_observations(datestring, deprecated, capfd):
     else:
         assert "deprecat" not in captured.err.lower()
         assert "deprecat" not in captured.out.lower()
+
+
+def test_observations(setup_case):
+    observations = EnkfObs.from_ert_config(
+        setup_case("simple_config", "minimum_config")
+    )
+    count = 10
+    summary_key = "test_key"
+    observation_key = "test_obs_key"
+    summary_observation_node = EnkfConfigNode.createSummaryConfigNode(
+        summary_key, LoadFailTypeEnum.LOAD_FAIL_EXIT
+    )
+    observation_vector = ObsVector(
+        EnkfObservationImplementationType.SUMMARY_OBS,
+        observation_key,
+        summary_observation_node,
+        count,
+    )
+
+    observations.addObservationVector(observation_vector)
+
+    values = []
+    for index in range(0, count):
+        value = index * 10.5
+        std = index / 10.0
+        summary_observation_node = SummaryObservation(
+            summary_key, observation_key, value, std
+        )
+        observation_vector.installNode(index, summary_observation_node)
+        assert observation_vector.getNode(index) == summary_observation_node
+        assert summary_observation_node.getValue() == value
+        values.append((index, value, std))
+
+    test_vector = observations[observation_key]
+    index = 0
+    for node in test_vector:
+        assert isinstance(node, SummaryObservation)
+        assert node.getValue() == index * 10.5
+        index += 1
+
+    assert observation_vector == test_vector
+    for index, value, std in values:
+        assert test_vector.isActive(index)
+
+        summary_observation_node = test_vector.getNode(index)
+
+        assert summary_observation_node.getValue() == value
+        assert summary_observation_node.getStandardDeviation() == std
+        assert summary_observation_node.getSummaryKey() == summary_key
